requirements.txt
streamlit
pandas
numpy
scikit-learn
finnhub-python
pandas-ta
